feat: better button types + misc changes
