+++ packages/addressBadge/AddressBadgeStyles.tsx
@@ -2,7 +2,9 @@ import { IdenticonBadge } from '@lidofinance/identicon'
 import { InferStyledComponentPropsWithoutRef } from '@lidofinance/utils'
 import styled from 'styled-components'
 
-export type AddressBadgeStyleProps = InferStyledComponentPropsWithoutRef<typeof AddressBadgeStyle>
+export type AddressBadgeStyleProps = InferStyledComponentPropsWithoutRef<
+  typeof AddressBadgeStyle
+>
 
 export const AddressBadgeStyle = styled(IdenticonBadge)`
   max-width: 100%;
