+++ packages/block/Block.stories.tsx
@@ -1,4 +1,3 @@
-import React from 'react'
 import { Story, Meta } from '@storybook/react'
 import { Block, BlockProps, BlockColor, BlockVariant } from '.'
