+++ packages/button/Button.test.tsx
@@ -1,4 +1,3 @@
-import React from 'react'
 import { render } from '@testing-library/react'
 import { ThemeProvider, themeDefault } from '@lidofinance/theme'
 import 'jest-styled-components'
