+++ packages/button/ButtonIcon.tsx
@@ -1,17 +1,19 @@
-import React, { ForwardedRef, forwardRef } from 'react'
+import { ForwardedRef, forwardRef } from 'react'
 import {
   ButtonWrapperStyle,
   ButtonIconStyle,
   ButtonContentStyle,
 } from './ButtonIconStyles'
-import { ButtonIconProps } from './types'
-import Button from './Button'
+import Button, { ButtonProps } from './Button'
+
+export type ButtonIconProps = {
+  icon: JSX.Element
+} & ButtonProps
 
 function ButtonIcon(
-  props: ButtonIconProps,
+  { icon, children, ...rest }: ButtonIconProps,
   ref?: ForwardedRef<HTMLButtonElement>
 ) {
-  const { icon, children, ...rest } = props
   const hasNoChildren = !children
 
   return (
